Update InclusionPayload to match Slasher contract specification
Changes InclusionPayload structure to align with on-chain contract:
- Remove tx_hash, nonce, gas_limit fields (now derivable from signed_tx)
- Remove InclusionMetadata struct (no longer needed)
- Rename tx_data -> signed_tx to match contract field name
- Add tx_hash() helper method for deriving transaction hash

Additionally:
- Migrate slot_congestion table from DECIMAL to DOUBLE PRECISION
- Remove all CAST operations in SQL queries (cleaner code)
- Update fee pricing engine to parse new payload structure
- Add sqlx query cache for offline builds
- All tests passing

This matches the updated Slasher contract where InclusionPayload contains:
  slot: uint64
  signed_tx: Bytes (full signed transaction)

+++ migrations/006_create_slot_congestion_table.sql
@@ -8,11 +8,11 @@ CREATE TABLE slot_congestion (
     -- Gas usage tracking
     preconfirmed_gas BIGINT NOT NULL DEFAULT 0,
     total_gas_limit BIGINT NOT NULL DEFAULT 30000000,
-    gas_used_ratio DECIMAL(5,4) NOT NULL DEFAULT 0.0000,
+    gas_used_ratio DOUBLE PRECISION NOT NULL DEFAULT 0.0,
 
     -- Fee calculation metadata
     base_gas_price BIGINT NOT NULL,
-    calculated_fee_multiplier DECIMAL(10,6) NOT NULL DEFAULT 1.0,
+    calculated_fee_multiplier DOUBLE PRECISION NOT NULL DEFAULT 1.0,
     current_tx_price BIGINT NOT NULL,
 
     -- Timestamp tracking
